requirements.txt
alembic==1.3.2
Babel==2.8.0
blinker==1.4
Click==7.0
dominate==2.4.0
elasticsearch==6.1.1
Flask==1.1.1
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
langdetect==1.0.7
langid==1.1.6
Mako==1.1.0
MarkupSafe==1.1.1
numpy==1.18.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.12
urllib3==1.22
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
